Tune robot parameters and fix rotation type for auto
Change rotation from int to double for finer angle control, increase
flywheel max RPM to 5000, tighten turret ready threshold, and use
separate red/blue OTOS starting positions.

+++ TeamCode/src/main/java/com/buddyram/rframe/ftc/DriveToAndRotateAction.java
@@ -15,12 +15,12 @@ public class DriveToAndRotateAction<T extends Navigatable<HolonomicDriveTrain>> 
     private final Vector3D target;
     private final double accuracy;
     private final CalculateDriveSpeed speed;
-    private final int rotation;
+    private final double rotation;
 
 
     // public DriveToAndRotateAction(Vector3D target, double accuracy, int rotation) {}
 
-    public DriveToAndRotateAction(Vector3D target, double accuracy, CalculateDriveSpeed speed, int rotation) {
+    public DriveToAndRotateAction(Vector3D target, double accuracy, CalculateDriveSpeed speed, double rotation) {
         this.target = target;
         this.accuracy = accuracy;
         this.speed = speed;
